% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLists.R
\name{binaryStacks}
\alias{binaryStacks}
\alias{binaryStack}
\alias{maskWPine}
\title{Create a set of binary stacks from quantitative stacks}
\usage{
binaryStacks(
  sims,
  thresholdAttackTreesMinDetectable = 1.4,
  thresholdPineProportion = 0.3,
  stackName = "predictedStack",
  propPineRasName = "propPineRas"
)

binaryStack(
  stk,
  propPineRas,
  thresholdAttackTreesMinDetectable = 1.4,
  thresholdPineProportion = 0.3
)

maskWPine(ras, propPineRas, thresholdPineProportion)
}
\arguments{
\item{sims}{a list of simLists}

\item{thresholdAttackTreesMinDetectable}{A scalar that delineates presence from
an absense. Default is 1.4, which was derived from initial efforts to find
a single value that works in all years.}

\item{thresholdPineProportion}{A scalar. Values on the propPineRas that are below
this threshold will be masked out, i.e., set to NA.}

\item{stackName}{Character string. The name of the stack inside the simLists to use}

\item{propPineRasName}{Character string. The name of the RasterLayer inside the simLists to use
for proportion Pine.}

\item{stk}{A stack of abundance}

\item{propPineRas}{A RasterLayer that has values between 0 and 1, representing
the proportion of pine cover in the pixel.}

\item{ras}{A RasterLayer of abundance}
}
\description{
This is for converting forecasted abundance stacks in a list of simList objects
into a list of binary stacks (presence/absence)

This is for converting forecasted abundance stack in a simList object
into a binary stack (presence/absence)

The mask will occur with values of NA or values of 0.
}
