% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cleanUpPredictionRas}
\alias{cleanUpPredictionRas}
\title{Cleans the predicted raster}
\usage{
cleanUpPredictionRas(
  rasLog,
  propPineRas,
  thresholdAttackTreesMinDetectable = 1.4,
  thresholdPineProportion = 0.3
)
}
\arguments{
\item{rasLog}{A RasterLayer of predicted mass attack, on a log scale}

\item{propPineRas}{A RasterLayer of proportion pine cover.}

\item{thresholdAttackTreesMinDetectable}{A scalar. This will have come from
an \code{optim} call to estimate this using AUC and ROC. THIS IS NOT ON A LOG
SCALE. It will be logged internally.}

\item{thresholdPineProportion}{A scalar. The proportion of pine}

\item{years}{A scalar indicating how many years are included in the \code{rasLog}}
}
\value{
A \code{RasterLayer} with ever
}
\description{
Cleans the predicted raster
}
