% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cumulativeMap}
\alias{cumulativeMap}
\title{Calculate cumulative sum of a stack, with optional log}
\usage{
cumulativeMap(stk, log = TRUE)
}
\arguments{
\item{stk}{a RasterStack}

\item{log}{Logical. Will take \code{log(map) + 1}}
}
\description{
Calculate cumulative sum of a stack, with optional log
}
