% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growthFunction}
\alias{growthFunction}
\title{MPB red top growth function}
\usage{
growthFunction(x, s, dataset, growthData)
}
\arguments{
\item{x}{beetle density}

\item{s}{climate suitability factor}

\item{dataset}{dataset name used for fitting:
one of `"Berryman1979_fit"`, `"Berryman1979_forced"`, `"Boone2011"`.}

\item{growthData}{MPB red top growth data with which to fit.}
}
\description{
From Cooke and Carroll (2017).
}
