% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{kernel_twoDT}
\alias{kernel_twoDT}
\alias{kernel_twoDT_mean}
\title{2Dt kernel}
\usage{
kernel_twoDT(dist, mu, p)

kernel_twoDT_mean(mu, m)
}
\arguments{
\item{dist}{A vector of distances}

\item{mu}{The first parameter of the 2Dt kernel. This represents about 0.9 of the
mean dispersal distance}

\item{p}{The second parameter of the 2Dt kernel. This changes the shape.}
}
\value{
A vector of probabilities
}
\description{
This is a 2-dimensional t-distribution that can be used for dispersal of natural
entities. It has many characteristics (see Clark et al 1999)
}
\references{
Clark, J. S., M. Silman, R. Kern, E. Macklin, and J. HilleRisLambers. 1999.
Seed dispersal near and far: patterns across temperate and tropical forests.
Ecology 80:1475–1494.
}
