% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyArea.R
\name{mpbStudyArea}
\alias{mpbStudyArea}
\title{Create study area based o ecoregion selection}
\usage{
mpbStudyArea(ecoregions = c(112, 120, 122, 124, 126), targetCRS, cPath, dPath)
}
\arguments{
\item{ecoregions}{numeric vector indicating which ecoregions to be included as part of the study
area. Derived from \url{http://sis.agr.gc.ca/cansis/nsdb/ecostrat/region/ecoregion_shp.zip}.}

\item{targetCRS}{target CRS string to use for reprojecting ecodistrict (study area) polygons.}

\item{cPath}{cache path}

\item{dPath}{destination path}
}
\value{
an \code{sf} object
}
\description{
Create study area based o ecoregion selection
}
