% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{xt}
\alias{xt}
\title{MPB recruitment function}
\usage{
xt(xtminus1, cs, dataset, massAttacksMap, growthData)
}
\arguments{
\item{xtminus1}{previous year's attack density}

\item{cs}{climate suitability factor}

\item{dataset}{dataset name used for fitting:
one of \code{"Berryman1979_fit"}, \code{"Berryman1979_forced"}, \code{"Boone2011"}.}

\item{massAttacksMap}{\code{RasterLayer} of red attack densities}

\item{growthData}{MPB red top growth data with which to fit.}
}
\description{
From Cooke & Carroll (2017).
}
